+++ .eslintrc
@@ -18,5 +18,6 @@
     "import/prefer-default-export": "off",
     "max-len": ["error", { "code": 80 }],
     "no-confusing-arrow": "error",
+    "import/no-extraneous-dependencies": "off", //?
   }
 }
