+++ src/components/app.scss
@@ -24,16 +24,51 @@ body {
       align-items: center;
 
       border-bottom: 1px solid #ccc;
+      height: 100%;
 
       a {
         color: #333;
         text-decoration: none;
+        height: 100%;
+        margin-bottom: 5px;
 
         .button {
-          margin: 5px;
-          padding: 2px 7px;
-          background: #ccc;
+          padding: 4px 10px;
+          background: #fff;
+          height: 100%;
+
+          border-left: 0px solid #333;
+          border-right: 0px solid #333;
+
+          display: flex;
+          align-items: center;
+          justify-content: space-between;
+          font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+            Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+          font-size: 18px;
+          font-weight: 700;
+
           
+          border: 1px solid #fff;
+          border-radius: 24px;
+          transition: 0.5s;
+
+          * {
+            margin-right: 6px;
+            fill: #333;
+            stroke: #333;
+            transition: 0.5s;
+          }
+        }
+
+        .button:hover {
+          
+          border: 1px solid #ccc;
+          border-radius: 2px;
+          * {
+            fill: #8bc34a;
+            stroke: green;
+          }
         }
       }
     }
