+++ src/components/cardDetail/cardDetail.jsx
@@ -2,24 +2,17 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import classNames from 'classnames';
 import colorLabel from 'styles';
-import styles from './profileCard';
+import styles from './cardDetail.scss';
 import { AddChecklist } from './addChecklist';
 import { Checklist } from './checklist';
 
-export const ProfileCard = ({ match: { params = {} } = {} }) => {
-  const { id: cardId = 'card 1' } = params;
-  const checklists = useSelector((state) => state.checklists);
-  const cards = useSelector((state) => state.cards);
-  const card = cards.filter(({ id }) => id === cardId)[0];
-  const labelsAll = useSelector((state) => state.labels);
-  const {
-    title, description, date, labels,
-  } = card;
+export const CardDetail = () => (
+  <div className={styles.profile}>
+    <div className="content">Detail</div>
+  </div>
+);
 
-  return (
-    <div className={styles.profile}>
-      <div className="content">
-        <div className="wrapper-title">
+/* <div className="wrapper-title">
           <div className="title">{title}</div>
           <div className="labels">
             {labelsAll
@@ -51,8 +44,4 @@ export const ProfileCard = ({ match: { params = {} } = {} }) => {
               title={checklist.title}
               checklistID={checklist.id}
             />
-          ))}
-      </div>
-    </div>
-  );
-};
+          ))} */
