+++ src/components/cardDetail/cardDetail.scss
