+++ src/components/cardDetail/index.js
@@ -1,0 +1,1 @@
+export { CardDetail } from './cardDetail';
