+++ src/components/main/board/board.scss
@@ -11,6 +11,7 @@
 
   .wrapper {
     background: #fff;
+    color: #333;
     width: 100%;
     border-color: #333;
   }
