+++ src/components/main/board/cardMini/cardMini.scss
@@ -3,12 +3,14 @@
   margin-bottom: 4px;
   display: flex;
   flex-direction: column;
-  flex: 1;
   overflow: hidden;
   border: 1px solid #333;
   border-radius: 4px;
+  margin: 4px 4px 0 4px;
 
   .wrapper-card {
+    background: #fff;
+    color: #333;
     width: 100%;
     display: flex;
     flex-direction: column;
@@ -17,8 +19,8 @@
     border-color: #fff;
 
     .wrapper-colorLabel {
-      background: inherit;
-      color: inherit;
+      background: #333;
+      color: #fff;
       padding: 2px;
       margin: 0;
       outline: none;
