+++ src/components/main/board/cardMini/index.js
@@ -1,0 +1,1 @@
+export { CardMini } from './cardMini';
