+++ src/helpers/addItemInput/addItemInput.scss
@@ -1,5 +1,6 @@
 .addItem :global {
   background: inherit;
+  color: inherit;
   display: flex;
   flex: 1;
   align-items: flex-start;
@@ -7,7 +8,7 @@
   width: 100%;
   position: relative;
   margin: 0;
-  padding: 0;
+  padding: 1px;
   border: 1px solid #000;
   border-color: #333;
   border-radius: 4px;
@@ -20,10 +21,10 @@
     background: inherit;
     color: inherit;
     margin: 0;
-    padding: 8px 56px 5px 8px;
     width: 100%;
-    min-height: 24px;
+    min-height: 28px;
     border: 0px solid #000;
+    border-radius: 4px;
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
       Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
     font-size: 16px;
@@ -34,28 +35,48 @@
     color: #eee;
   }
 
-  .board {
+  .newBoard {
     background: #fff;
     color: inherit;
+    padding-right: 32px;
+    padding: 8px 32px 5px 8px;
+    cursor: default;
+  }
+  .newBoard:focus {
+    cursor: text;
+  }
+
+  .boardTitle {
+    background: #fff;
+    color: inherit;
+    cursor: default;
   }
 
   .button {
     position: absolute;
     right: 0;
     top: 0;
-    background: #8bc34a;
+    background: inherit;
     color: #eee;
     margin: 0;
     padding: 0;
-    border: 1px solid #eee;
-    border-radius: 4px;
+    border: 1px solid #fff;
+    border-radius: 0 4px;
     outline: none;
-    height: 100%;
-    min-width: 48px;
     display: flex;
-    align-items: center;
-    justify-content: center;
+    justify-content: flex-end;
+    align-items: flex-start;
     font-size: 24px;
     font-weight: 700;
+
+    * {
+      height: 28px;
+      width: 28px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      fill: #8bc34a;
+      border-radius: 4px;
+    }
   }
 }
